add auto-resize on display change (finally!)
When the OS sends a notification that the display changes, unless
specific size-controlling options were specified, adjust the window
and display to match the new screen size.

In the old days, this was hopeless, because text metrics would change
with resolution. I think we're past all that with the modern drawing
library, though.

It still won't work in all cases. It's not compatible with
`size-in-pixels`, but no one uses that anymore. It won't work with the
`slideshow-repl` package.