Implement nitem
These changes adds the implementation for nitem, which is similar to item except
that it prints numbers instead of bullets. This will automatically increment the
count for each slide and on every new slide the count will be reset.

There is a default separator that comes after the number "." that can be
changed. Also, the number itself can be changed, but the count for the slide
will still be incremented if this is provided.